+++ examples/issue68.rs
@@ -3,6 +3,7 @@
 use quick_xml::events::Event;
 use quick_xml::name::Namespace;
 use quick_xml::Reader;
+use std::convert::TryFrom;
 use std::io::Read;
 
 struct Resource {
@@ -83,7 +84,9 @@ fn parse_report(xml_data: &str) -> Vec<Resource> {
     loop {
         match reader.read_namespaced_event(&mut buf, &mut ns_buffer) {
             Ok((namespace_value, Event::Start(e))) => {
-                let namespace_value = namespace_value.unwrap_or(Namespace(b""));
+                let namespace_value = Option::<Namespace>::try_from(namespace_value)
+                    .unwrap_or_default() // Treat unknown prefixes as not bound to any namespace
+                    .unwrap_or(Namespace(b""));
                 match (depth, state, namespace_value.as_ref(), e.local_name()) {
                     (0, State::Root, b"DAV:", b"multistatus") => state = State::MultiStatus,
                     (1, State::MultiStatus, b"DAV:", b"response") => {
@@ -98,7 +101,9 @@ fn parse_report(xml_data: &str) -> Vec<Resource> {
                 depth += 1;
             }
             Ok((namespace_value, Event::End(e))) => {
-                let namespace_value = namespace_value.unwrap_or(Namespace(b""));
+                let namespace_value = Option::<Namespace>::try_from(namespace_value)
+                    .unwrap_or_default() // Treat unknown prefixes as not bound to any namespace
+                    .unwrap_or(Namespace(b""));
                 let local_name = e.local_name();
                 match (depth, state, namespace_value.as_ref(), local_name) {
                     (1, State::MultiStatus, b"DAV:", b"multistatus") => state = State::Root,