+++ tests/xmlrs_reader_tests.rs
@@ -1,5 +1,5 @@
 use quick_xml::events::{BytesStart, Event};
-use quick_xml::name::Namespace;
+use quick_xml::name::ResolveResult;
 use quick_xml::{Reader, Result};
 use std::borrow::Cow;
 use std::str::from_utf8;
@@ -388,7 +388,7 @@ fn test_bytes(input: &[u8], output: &[u8], is_short: bool) {
     loop {
         buf.clear();
         let event = reader.read_namespaced_event(&mut buf, &mut ns_buffer);
-        let line = xmlrs_display(&event, &reader);
+        let line = xmlrs_display(event, &reader);
         if let Some((n, spec)) = spec_lines.next() {
             if spec.trim() == "EndDocument" {
                 break;
@@ -427,15 +427,15 @@ fn test_bytes(input: &[u8], output: &[u8], is_short: bool) {
     }
 }
 
-fn namespace_name(n: &Option<Namespace>, name: &[u8]) -> String {
+fn namespace_name(n: ResolveResult, name: &[u8]) -> String {
     match n {
         // Produces string '{namespace}local_name'
-        Some(n) => format!(
+        ResolveResult::Bound(n) => format!(
             "{{{}}}{}",
             from_utf8(n.as_ref()).unwrap(),
             from_utf8(name).unwrap()
         ),
-        None => from_utf8(name).unwrap().to_owned(),
+        _ => from_utf8(name).unwrap().to_owned(),
     }
 }
 
@@ -469,9 +469,9 @@ fn decode<'a>(text: &'a [u8], reader: &Reader<&[u8]>) -> Cow<'a, str> {
     decoded
 }
 
-fn xmlrs_display(opt_event: &Result<(Option<Namespace>, Event)>, reader: &Reader<&[u8]>) -> String {
+fn xmlrs_display(opt_event: Result<(ResolveResult, Event)>, reader: &Reader<&[u8]>) -> String {
     match opt_event {
-        Ok((ref n, Event::Start(ref e))) => {
+        Ok((n, Event::Start(ref e))) => {
             let name = namespace_name(n, decode(e.name(), reader).as_bytes());
             match make_attrs(e) {
                 Ok(ref attrs) if attrs.is_empty() => format!("StartElement({})", &name),
@@ -479,7 +479,7 @@ fn xmlrs_display(opt_event: &Result<(Option<Namespace>, Event)>, reader: &Reader
                 Err(e) => format!("StartElement({}, attr-error: {})", &name, &e),
             }
         }
-        Ok((ref n, Event::Empty(ref e))) => {
+        Ok((n, Event::Empty(ref e))) => {
             let name = namespace_name(n, decode(e.name(), reader).as_bytes());
             match make_attrs(e) {
                 Ok(ref attrs) if attrs.is_empty() => format!("EmptyElement({})", &name),
@@ -487,7 +487,7 @@ fn xmlrs_display(opt_event: &Result<(Option<Namespace>, Event)>, reader: &Reader
                 Err(e) => format!("EmptyElement({}, attr-error: {})", &name, &e),
             }
         }
-        Ok((ref n, Event::End(ref e))) => {
+        Ok((n, Event::End(ref e))) => {
             let name = namespace_name(n, decode(e.name(), reader).as_bytes());
             format!("EndElement({})", name)
         }